Remove the need for plugins and core to match versions
This was made possible by breaking changes made to the core.

Specifically, we reverted the previous change that required specifying a distinct `main` function in the metadata and instead went back to using _prolog as the entrypoint. Additionally, we now pass an API object that plugins will use to communicate with the core to remove the need to link against sy_core directly.